Move loading logic into a custom hook
